Add files via upload
In this project, I have worked on making a Python Script be run on three auction links for the book "Yoga Anatomy" to grab the updated auction price or a simple price change. 
I have used the library "Beautiful Soup" of Python to scrape as well as parse the data.